Bug 77999 - Part 1: Add nsDocumentRuleResultCacheKey.
nsDocumentRuleResultCacheKey is analagous to nsMediaQueryResultCacheKey
and stores the result of matching the @-moz-document rules from a set
of style sheets.  In a later patch, nsCSSRuleProcessor will build up
an nsDocumentRuleResultCacheKey as it visits the @-moz-document rules
while building its RuleCascadeData.

Rather than represent the result using a list of both the matching and
non-matching rules, we just store the matched rules.  The assumption is
that in situations where we have a large number of rules -- such as the
thousands added by AdBlock Plus -- that only a small number will be
matched.  Thus to check if the nsDocumentRuleResultCacheKey matches a
given nsPresContext, we also need the entire list of @-moz-document
rules to know which rules must not match.