Bug 77999 - Part 2: Add RuleProcessorCache.
The RuleProcessorCache is a singleton object that caches
nsCSSRuleProcessors keyed off a list of style sheets and the result of
evaluating all @-moz-documents in the style sheets.  In a later patch,
nsStyleSet will get and put nsCSSRuleProcessors from/to the
RuleProcessorCache.

We add some state bits to CSSStyleSheet and nsCSSRuleProcessor to track
whether they are in the RuleProcessorCache.  This lets us remove them
from the RuleProcessorCache when they're going away.