Bug 77999 - Part 3: Gather document rules and produce an nsDocumentRuleResultCacheKey in nsCSSRuleProcessors.
The first time we build a RuleCascadeData for a shared rule processor,
we build up its nsDocumentRuleResultCacheKey.  Note that we also
traverse into failing @media, @supports and @-moz-document rules, which
means we might look at the result of some @-moz-document rules that
don't actually affect whether a cache lookup is a hit or a miss.  The
presence of @-moz-document rules inside @media etc. rules should be
rare, and looking at all of them in the sheets lets us avoid the
complication of having different document cache key results for
different media.